pyproject.toml
[project]
name = "gerador-eduten"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.7",
    "numpy>=2.4.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "streamlit>=1.55.0",
    "xlrd>=2.0.2",
]
